mm-video: vidc: Restore sanity representing buffer sizes in dyn buf mode
nFilledLen and nAllocLen are buffer handle sizes in dynamic buffer mode.
This led to obnoxious checks in multiple places to account for dynamic
buffer mode.  Furthermore it led to basic sanity checks such as
`nFilledLen + nOffset < nAllocLen` failing.

To get around this, overwrite the OMX_BUFFERHEADER's nFilledLen /
nAllocLen parameters on FTB and restore them on FBD.  This localises the
need for the aforementioned irritating check.

Change-Id: Iac96d3b680d7e444e04680fdcf6a49913e5c5d0b
CRs-Fixed: 647383